refactor: use unified sorting
