+++ dash/src/components/TeamRadios.tsx
@@ -4,7 +4,7 @@ import clsx from "clsx";
 
 import { useDataStore } from "@/stores/useDataStore";
 
-import { sortUtc } from "@/lib/sorting/sortUtc";
+import { sortUtc } from "@/lib/sorting";
 
 import TeamRadioMessage from "@/components/TeamRadioMessage";
